+++ http4k-format-argo/src/test/kotlin/org/http4k/format/ArgoTest.kt
@@ -3,6 +3,10 @@ package org.http4k.format
 import argo.jdom.JsonNode
 import argo.jdom.JsonRootNode
 
-class ArgoTest : JsonContract<JsonRootNode, JsonNode>(Argo)
+class ArgoTest : JsonContract<JsonRootNode, JsonNode>(Argo) {
+    override val prettyString = """{
+	"hello": "world"
+}"""
+}
 class ArgoJsonErrorResponseRendererContractTest : JsonErrorResponseRendererContract<JsonRootNode, JsonNode>(Argo)
 class ArgoGenerateDataClassesTest : GenerateDataClassesContract<JsonRootNode, JsonNode>(Argo)