+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This list is not currently intended to be all-encompassing - it will document major and breaking API changes with their rationale when appropriate:
 
+### v3.9.0
+- Added compactify and prettify to Json implementations
+- Added `Json.hasBody` Hamkrest matchers for comparing bodies. Note these are extension methods and need to be referenced/imported as such.
+
 ### v3.8.0
 - Added facility for non-blocking websocket client to react to onConnect event. This API is
 the same as the inbound, server-side API - ie. there are no explicit connection event handlers. H/T @tom for the idea.